Adding a couple other SCALE 13x talks
More talks:

https://www.socallinuxexpo.org/scale/13x/presentations/usb-embedded-systems-quirky-versatile
and
https://www.socallinuxexpo.org/scale/13x/presentations/tale-two-build-systems-buildroot-and-openembedded

These slides were made in inkscape with jessyink extension

*_paths.svg files under bin/ are the versions where all fonts
were converted to paths so they would look right regardless of fonts
installed on the system (wasn't sure how to save an inkscape file
with fonts embedded, though perhaps that is possible). Those are
the ones that were used at SCALE, with the exception of one typo
discovered _during_ the buildroot/OE talk being fixed.

other .svg files are the original versions with fonts,
to be used for editing. Note that the jessyink effects that
fade objects in and out need to be redone after converting
fonts to paths, so if any significant editing needs done it
would be best to look into embedding fonts instead...

usb_quirky_versatile_outline.txt is the original outline for the
USB talk, basis of an earlier LUG talk on the same topic.